[#noissue] Refactor TagUtils to move to util package and deprecate old implementation
