+++ metric-module/metric/src/main/java/com/navercorp/pinpoint/metric/web/mapping/Field.java
@@ -19,7 +19,7 @@ package com.navercorp.pinpoint.metric.web.mapping;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.navercorp.pinpoint.metric.common.model.Tag;
-import com.navercorp.pinpoint.metric.common.model.TagUtils;
+import com.navercorp.pinpoint.metric.common.util.TagUtils;
 import com.navercorp.pinpoint.metric.web.model.basic.metric.group.MatchingRule;
 
 import java.util.List;
